refactor: move vars creation to own file
